pyproject.toml
[project]
name = "population-decline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apache-airflow>=3.1.3",
    "minio>=7.2.20",
    "mlflow>=3.6.0",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
]
